requirements.txt
streamlit>=1.32.0
google-genai>=1.14.0
Pillow>=10.0.0
anthropic>=0.40.0
PyJWT>=2.8.0
requests>=2.31.0
moviepy>=1.0.3
imageio-ffmpeg>=0.4.8
numpy>=1.24.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
